+++ plugins/scaffolder-frontend-workflow/src/components/ModalWorkflow/ModalWorkflow.tsx
@@ -1,19 +1,24 @@
-import React, { ReactNode, useCallback, useState } from 'react';
+import React, { ReactNode, useCallback, useEffect, useState } from 'react';
 import { type WorkflowProps } from '@backstage/plugin-scaffolder-react/alpha';
 import { Modal, ModalProps } from '../Modal/Modal';
 import { Workflow } from '../Form/Workflow';
+import { stringifyEntityRef } from '@backstage/catalog-model';
+import { useRunWorkflow, useWorkflowManifest } from '../../hooks';
 
 type ModalWorkflowProps = Pick<
   WorkflowProps,
   'namespace' | 'templateName' | 'onCreate' | 'initialState'
-> & Pick<ModalProps, 'onClick' | 'onClose' | 'open'> & {
-  children: ReactNode;
-  onComplete?: () => void;
-  onError?: (e: Error) => void;
-  className?: string;
-};
+> &
+  Pick<ModalProps, 'onClick' | 'onClose' | 'open'> & {
+    children: ReactNode;
+    onComplete?: () => void;
+    onError?: (e: Error) => void;
+    className?: string;
+  };
 
 export function ModalWorkflow({
+  namespace,
+  templateName,
   onComplete,
   onError,
   className,
@@ -21,34 +26,58 @@ export function ModalWorkflow({
   onClick,
   onClose,
   ...props
-}: ModalWorkflowProps): JSX.Element {
+}: ModalWorkflowProps): JSX.Element | null {
   const [modalExpanded, setModalExpanded] = useState(false);
-  const onTaskComplete = useCallback(async () => {
-    setModalExpanded(false);
 
+  const templateRef = stringifyEntityRef({
+    kind: 'Template',
+    namespace: namespace,
+    name: templateName,
+  });
+
+  const { loading, manifest } = useWorkflowManifest({
+    name: templateName,
+    namespace: namespace,
+  });
+
+  const onTaskComplete = useCallback(async () => {
     onComplete?.();
   }, [onComplete]);
 
-  const onCloseHandler = useCallback((e) => {
-    setModalExpanded(false);
+  const onCloseHandler = useCallback(
+    e => {
+      setModalExpanded(false);
 
-    if (onClose) {
-      onClose(e, "backdropClick");
+      if (onClose) {
+        onClose(e, 'backdropClick');
+      }
+    },
+    [onClose],
+  );
+
+  const workflow = useRunWorkflow({
+    templateRef,
+    onComplete: onTaskComplete,
+  });
+
+  useEffect(() => {
+    if (workflow.taskStream.loading === false) {
+      setModalExpanded(true);
     }
-  }, [onClose]);
+  }, [workflow.taskStream.loading]);
 
-  return (
+  if (loading) {
+    return <>Loading template...</>;
+  }
+
+  return manifest ? (
     <Modal
       open={open}
       onClick={onClick}
       onClose={onCloseHandler}
       fullyExpanded={modalExpanded}
     >
-      <Workflow
-        {...props}
-        onTaskCreated={setModalExpanded}
-        onComplete={onTaskComplete}
-      />
+      <Workflow {...props} manifest={manifest} workflow={workflow} />
     </Modal>
-  );
+  ) : null;
 }
